requirements.txt
djangorestframework
markdown    
django-filter
django-extensions 
#pygraphviz
django-countries
django-cors-headers
djangorestframework-simplejwt
social-auth-app-django
django-allauth
inflection
uritemplate
drf-spectacular
django-tailwind
django-tailwind[reload]
Pillow
pandas
pymysql>=1.1.0
psycopg2-binary>=2.9.9 
pymysql
django-admin-interface
python-webpack-boilerplate
